Significant changes to clean up Link interface
