+++ dmpipe/dm_plotting.py
@@ -80,38 +80,6 @@ class PlotDMSpectra(Link):
             dm_plot_by_chan[0].savefig(args.outfile.replace('.png', '_%1.FGeV.png'%args.mass))
 
 
-class PlotCastro(Link):
-    """Small class wrap an analysis script.
-    
-    This is useful for parallelizing analysis using the fermipy.jobs module.
-    """
-    appname = 'dmpipe-plot-castro'
-    linkname_default = 'plot-castro'
-    usage = '%s [options]'%(appname)
-    description = "Plot likelihood v. flux normalization and energy"
-
-    default_options = dict(infile=defaults.generic['infile'],
-                           outfile=defaults.generic['outfile'])
-
-    def __init__(self, **kwargs):
-        """C'tor
-        """
-        linkname, init_dict = self._init_dict(**kwargs)
-        super(PlotCastro, self).__init__(linkname, **init_dict)
-
-    def run_analysis(self, argv):
-        """Run this analysis"""
-        args = self._parser.parse_args(argv)
-        castro_data = CastroData.create_from_sedfile(args.infile)
-        ylims = [1e-8, 1e-5]
-        
-        plot = plotCastro(castro_data, ylims)
-        if args.outfile:
-            plot[0].savefig(args.outfile)   
-            return None
-        return plot
-  
-
 class PlotLimits(Link):
     """Small class wrap an analysis script.
     
@@ -209,60 +177,6 @@ class PlotDM(Link):
             dm_plot[0].savefig(args.outfile)
             return None
         return dm_plot
-
-
- 
-class PlotCastro_SG(ConfigMaker):
-    """Small class to generate configurations for this script
-
-    This adds the following arguments:
-    """
-    appname = 'dmpipe-plot-castro-sg'
-    usage = "%s [options]" % (appname)
-    description = "Make castro plots for set of targets"
-    clientclass = PlotCastro
-    
-    job_time = 60
-
-    default_options = dict(ttype=defaults.common['ttype'],
-                           targetlist=defaults.common['targetlist'],
-                           dry_run=defaults.common['dry_run'])
-
-    def __init__(self, link, **kwargs):
-        """C'tor
-        """
-        super(PlotCastro_SG, self).__init__(link,
-                                            options=kwargs.get('options',
-                                                               self.default_options.copy()))
-
-    def build_job_configs(self, args):
-        """Hook to build job configurations
-        """
-        job_configs = {}
-
-        ttype = args['ttype']
-        (targets_yaml, sim) = NAME_FACTORY.resolve_targetfile(args)
-        if targets_yaml is None:
-            return job_configs
-
-        targets = load_yaml(targets_yaml)
-
-        for target_name, target_list in targets.items():
-            for targ_prof in target_list:
-                name_keys = dict(target_type=ttype,
-                                 target_name=target_name,
-                                 profile=targ_prof,
-                                 fullpath=True)
-                targ_key = "%s_%s"%(target_name, targ_prof)
-                input_path = NAME_FACTORY.sedfile(**name_keys)
-                output_path = input_path.replace('.fits', '.png')
-                logfile = make_nfs_path(input_path.replace('.fits', '.log'))
-                job_config = dict(infile=input_path,
-                                  outfile=output_path,
-                                  logfile=logfile)
-                job_configs[targ_key] = job_config
-                
-        return job_configs
 
 
 class PlotLimits_SG(ConfigMaker):
@@ -558,8 +472,6 @@ class PlotStackedDM_SG(ConfigMaker):
 
 def register_classes():
     PlotDMSpectra.register_class()
-    PlotCastro.register_class()
-    PlotCastro_SG.register_class()
     PlotLimits.register_class()
     PlotLimits_SG.register_class()
     PlotDM.register_class()
