+++ dmpipe/name_policy.py
@@ -6,30 +6,18 @@ from __future__ import absolute_import, division, print_function
 
 import yaml
 from fermipy.jobs.utils import is_null, is_not_null
+from fermipy.jobs.name_policy import NameFactory as NameFactory_Base
 
-
-class NameFactory(object):
+class NameFactory(NameFactory_Base):
     """ Helper class to define file names and keys consistently. """
-
-    # Input configuration file
-    ttypeconfig_format = 'config/config_{target_type}.yaml'
 
     # DM spectral file target configruation
     specconfig_format = 'config/dm_spectra_{target_type}.yaml'
-
-    # DM random file target configruation
-    randconfig_format = 'config/random_{target_type}.yaml'
 
     # DM spectral file
     specfile_format = 'dm_spectra_{target_type}.fits'
 
     # target keys, these are how we specify various files associated with particular targets
-
-    # Directory for a particular target
-    targetdir_format = '{target_type}/{target_name}'
-
-    # Directory for simulations for a particular target
-    sim_targetdir_format = '{target_type}_sim/sim_{sim_name}/{target_name}'
 
     # Roster list file format
     rosterfile_format = '{target_type}/{rosterlist}'
@@ -37,26 +25,11 @@ class NameFactory(object):
     # Simulated rosterlist  file format
     sim_rosterfile_format = '{target_type}_sim/sim_{sim_name}/{rosterlist}'
 
-    # Targetlist file format
-    targetfile_format = '{target_type}/{targetlist}'
-
-    # Roster file format
-    sim_targetfile_format = '{target_type}_sim/sim_{sim_name}/{targetlist}'
-
-    # Information about a particular target profile
-    profilefile_format = '{target_type}/{target_name}/profile_{profile}.yaml'
-
-    # Information about a particular target profile
-    sim_profilefile_format = '{target_type}_sim/sim_{sim_name}/{target_name}/profile_{profile}.yaml'
-
     # Information about a particular target j factor
     j_valuefile_format = '{target_type}/{target_name}/j_val_{profile}.yaml'
 
     # Information about a particular target j factor
     sim_j_valuefile_format = '{target_type}_sim/sim_{sim_name}/{target_name}/j_val_{profile}.yaml'
-
-    # SED file for a particular target
-    sedfile_format = '{target_type}/{target_name}/sed_{profile}.fits'
 
     # DM likelilood file for a particular target (and j-factor prior)
     dmlikefile_format = '{target_type}/{target_name}/dmlike_{profile}_{jprior}.fits'
@@ -70,9 +43,6 @@ class NameFactory(object):
     # Stacked DM limits file for a particular roster (and j-factor prior)
     stackedlimitsfile_format = '{target_type}/stacked/limits_{roster_name}_{jprior}.fits'
 
-    # Simulated SED file for a particular target
-    sim_sedfile_format = '{target_type}_sim/sim_{sim_name}/{target_name}/sed_{profile}_{seed}.fits'
-
     # Simulated DM likelilood file for a particular target (and j-factor prior)
     sim_dmlikefile_format = '{target_type}_sim/sim_{sim_name}/{target_name}/dmlike_{profile}_{jprior}_{seed}.fits'
 
@@ -85,62 +55,21 @@ class NameFactory(object):
     # Stacked DM limits file for a particular roster (and j-factor prior)
     sim_stackedlimitsfile_format = '{target_type}_sim/sim_{sim_name}/stacked/limits_{roster_name}_{jprior}_{seed}.fits'
 
-    # Stamp files from scatter gather jobs
-    stamp_format = 'stamps/{linkname}.stamp'
-
-    # Full filepath
-    fullpath_format = '{basedir}/{localpath}'
-
     def __init__(self, **kwargs):
         """ C'tor.  Set baseline dictionary used to resolve names
         """
-        self.base_dict = kwargs.copy()
-
-    def update_base_dict(self, yamlfile):
-        """ Update the values in baseline dictionary used to resolve names
-        """
-        self.base_dict.update(**yaml.safe_load(open(yamlfile)))
-
-    def _format_from_dict(self, format_string, **kwargs):
-        """ Return a formatted file name dictionary components """
-        kwargs_copy = self.base_dict.copy()
-        kwargs_copy.update(**kwargs)
-        localpath = format_string.format(**kwargs_copy)
-        if kwargs.get('fullpath', False):
-            return self.fullpath(localpath=localpath)
-        else:
-            return localpath
-
-    def ttypeconfig(self, **kwargs):
-        """ return the name of the input configuration file
-        """
-        return self._format_from_dict(NameFactory.ttypeconfig_format, **kwargs)
+        super(NameFactory, self).__init__(**kwargs)
 
     def specconfig(self, **kwargs):
         """ return the name of the input configuration file
         """
         return self._format_from_dict(NameFactory.specconfig_format, **kwargs)
 
-    def randconfig(self, **kwargs):
-        """ return the name of the random direction configuration file
-        """
-        return self._format_from_dict(NameFactory.randconfig_format, **kwargs)    
-    
     def specfile(self, **kwargs):
         """ return the name of DM spectral file
         """
         return self._format_from_dict(NameFactory.specfile_format, **kwargs)
       
-    def targetdir(self, **kwargs):
-        """ return the name for the directory for a particular target
-        """
-        return self._format_from_dict(NameFactory.targetdir_format, **kwargs)
-    
-    def sim_targetdir(self, **kwargs):
-        """ return the name for the directory for a particular target
-        """
-        return self._format_from_dict(NameFactory.sim_targetdir_format, **kwargs)
-
     def rosterfile(self, **kwargs):
         """ return the name for the Roster list file
         """
@@ -151,26 +80,6 @@ class NameFactory(object):
         """
         return self._format_from_dict(NameFactory.sim_rosterfile_format, **kwargs)
     
-    def targetfile(self, **kwargs):
-        """ return the name for the Target list file
-        """
-        return self._format_from_dict(NameFactory.targetfile_format, **kwargs)
-    
-    def sim_targetfile(self, **kwargs):
-        """ return the name for the Target list file for simulation
-        """
-        return self._format_from_dict(NameFactory.sim_targetfile_format, **kwargs)
-    
-    def profilefile(self, **kwargs):
-        """ return the name of the yaml file with information about a partiuclar profile
-        """
-        return self._format_from_dict(NameFactory.profilefile_format, **kwargs)
-
-    def sim_profilefile(self, **kwargs):
-        """ return the name of the yaml file with information about a partiuclar profile
-        """
-        return self._format_from_dict(NameFactory.sim_profilefile_format, **kwargs)
-
     def j_valuefile(self, **kwargs):
         """ return the name of the yaml file with information about a partiuclar target j factor
         """
@@ -180,11 +89,6 @@ class NameFactory(object):
         """ return the name of the yaml file with information about a partiuclar target j factor
         """
         return self._format_from_dict(NameFactory.sim_j_valuefile_format, **kwargs)
-
-    def sedfile(self, **kwargs):
-        """ return the name for the SED file for a particular target
-        """
-        return self._format_from_dict(NameFactory.sedfile_format, **kwargs)
     
     def dmlikefile(self, **kwargs):
         """ return the name for the DM likelilood file for a particular target
@@ -205,14 +109,7 @@ class NameFactory(object):
         """ return the name for the stacked limits file for a particular roster
         """
         return self._format_from_dict(NameFactory.stackedlimitsfile_format, **kwargs)
-    
-    def sim_sedfile(self, **kwargs):
-        """ return the name for the simulated SED file for a particular target
-        """
-        if not kwargs.has_key('seed'):
-            kwargs['seed'] = 'SEED'
-        return self._format_from_dict(NameFactory.sim_sedfile_format, **kwargs)
-    
+        
     def sim_dmlikefile(self, **kwargs):
         """ return the name for the simulated DM likelilood file for a particular target
         """
@@ -240,51 +137,7 @@ class NameFactory(object):
         if not kwargs.has_key('seed'):
             kwargs['seed'] = 'SEED'
         return self._format_from_dict(NameFactory.sim_stackedlimitsfile_format, **kwargs)
-   
-    def stamp(self, **kwargs):
-        """Return the path for a stamp file for a scatter gather job"""
-        kwargs_copy = self.base_dict.copy()
-        kwargs_copy.update(**kwargs)
-        return NameFactory.stamp_format.format(**kwargs_copy)
-
-    def fullpath(self, **kwargs):
-        """Return a full path name for a given file
-        """
-        kwargs_copy = self.base_dict.copy()
-        kwargs_copy.update(**kwargs)
-        return NameFactory.fullpath_format.format(**kwargs_copy)
-
     
-    def resolve_targetfile(self, args, require_sim_name=False):
-        """Get the name of the targetfile based on the job arguments"""
-        ttype = args.get('ttype')
-        if is_null(ttype):
-            sys.stderr.write('Target type must be specified')
-            return (None, None)
-      
-        sim = args.get('sim')
-        if is_null(sim):
-            if require_sim_name:
-                sys.stderr.write('Simulation scenario must be specified')
-                return (None, None)
-            else:
-                sim = None
-
-        name_keys = dict(target_type=ttype,
-                         targetlist='target_list.yaml',
-                         sim_name=sim,
-                         fullpath=True)
-        if sim is None:
-            targetfile = self.targetfile(**name_keys)
-        else:
-            targetfile = self.sim_targetfile(**name_keys)
-
-        targets_override = args.get('targetfile')
-        if is_not_null(targets_override):
-            targetfile = targets_override
-        
-        return (targetfile, sim)
-  
     def resolve_rosterfile(self, args, require_sim_name=False):
         """Get the name of the roster based on the job arguments"""
         ttype = args.get('ttype')
@@ -343,16 +196,3 @@ class NameFactory(object):
             specconfig = spec_override
         return specconfig
 
-    def resolve_randconfig(self, args):
-        """Get the name of the specturm file based on the job arguments"""
-        ttype = args.get('ttype')
-        if is_null(ttype):
-            sys.stderr.write('Target type must be specified')
-            return None
-        name_keys = dict(target_type=ttype,
-                         fullpath=True)
-        randconfig = self.randconfig(**name_keys)
-        rand_override = args.get('rand_config')
-        if is_not_null(rand_override):
-            randconfig = rand_override
-        return randconfig
