+++ setup.py
@@ -25,23 +25,11 @@ setup(
     ],
     scripts=[],
     entry_points={'console_scripts': [
-        'dmpipe-analyze-roi = dmpipe.target_analysis:AnalyzeROI.main',
-        'dmpipe-analyze-roi-sg = dmpipe.target_analysis:AnalyzeROI_SG.main',
-        'dmpipe-analyze-sed = dmpipe.target_analysis:AnalyzeSED.main',
-        'dmpipe-analyze-sed-sg = dmpipe.target_analysis:AnalyzeSED_SG.main',
-        'dmpipe-prepare-targets = dmpipe.target_analysis:PrepareTargets.main',
-        'dmpipe-copy-base-roi = dmpipe.target_sim:CopyBaseROI.main',
-        'dmpipe-copy-base-roi-sg = dmpipe.target_sim:CopyBaseROI_SG.main',
-        'dmpipe-random-dir-gen = dmpipe.target_sim:RandomDirGen.main',
-        'dmpipe-random-dir-gen-sg = dmpipe.target_sim:RandomDirGen_SG.main',
-        'dmpipe-simulate-roi = dmpipe.target_sim:SimulateROI.main',
-        'dmpipe-simulate-roi-sg = dmpipe.target_sim:SimulateROI_SG.main',
+        'dmpipe-prepare-targets = dmpipe.dm_prepare:PrepareTargets.main',
         'dmpipe-convert-castro = dmpipe.dm_spectral:ConvertCastro.main',
         'dmpipe-convert-castro-sg = dmpipe.dm_spectral:ConvertCastro_SG.main',
         'dmpipe-stack-likelihood = dmpipe.dm_spectral:StackLikelihood.main',
         'dmpipe-stack-likelihood-sg = dmpipe.dm_spectral:StackLikelihood_SG.main',
-        'dmpipe-collect-sed = dmpipe.dm_collect:CollectSED.main',
-        'dmpipe-collect-sed-sg = dmpipe.dm_collect:CollectSED_SG.main',
         'dmpipe-collect-limits = dmpipe.dm_collect:CollectLimits.main',
         'dmpipe-collect-limits-sg = dmpipe.dm_collect:CollectLimits_SG.main',
         'dmpipe-collect-stacked-limits-sg = dmpipe.dm_collect:CollectStackedLimits_SG.main',
@@ -54,8 +42,6 @@ setup(
         'dmpipe-plot-limits = dmpipe.dm_plotting:PlotLimits.main', 
         'dmpipe-plot-limits-sg = dmpipe.dm_plotting:PlotLimits_SG.main', 
         'dmpipe-plot-stacked-limits-sg = dmpipe.dm_plotting:PlotStackedLimits_SG.main', 
-        'dmpipe-plot-castro = dmpipe.dm_plotting:PlotCastro.main',
-        'dmpipe-plot-castro-sg = dmpipe.dm_plotting:PlotCastro_SG.main',
         'dmpipe-calc-sed-limits = dmpipe.scripts.calc_sed_limits:main', 
     ]},
     install_requires=[
